Fix for Communicate flyout button.
Remove erroneously linked control variable.

The 'value' of a LLFlyoutButton is the (last) selected value:
a string. Linking it to some random boolean control variable
causes this value to be (re)assigned to the flyout button list
through LLControlVariable::setValue (whenever said boolean
changes, in this case when the floater is made to appear)
which calls getComparableValue with the string value (ie "mute list")
and tries to convert that to a boolean. Obviously That fails
and the value is set to "", effectively having nothing selected
anymore.

This fixes
http://code.google.com/p/singularity-viewer/issues/detail?id=21