Allow to LD_PRELOAD (any) library just for singularity-do-not-run-directly
This adds support for running:

AI_PRELOAD="/path/to/libfoo.so" ./singularity

Note that setting and exporting LD_PRELOAD there doesn't work,
because then those libs are loaded for every executable that
is being run after, including the plugins.

This patch also removes the use of tcmalloc from the plugins...
Before it exported LD_PRELOAD, using tcmalloc for all the command
line tools up till and including running the singularity binary,
as well as for the plugins. Now, it only preloads the tcmalloc
libraries for singularity-do-not-run-directly.