Comment out old Google Translate support and as such remove the need for JSON.
Note that on standalone, json headers are installed in json/,
no idea why in the packages this isn't the case... But anyway,
it isn't used at all anymore now so it doesn't really matter.

I didn't remove the json package for non-standalone and the
downloading thereof... (I can't test non-standalone).