Do not attach cam to avatar in Freeze Time mode.
We do, and the avatar moves then the cam jumps
away from the frozen avatar to where it really is.

Unfortunately, it's still not possible to attach
the cam to moving objects with this patch.
Also, scripted changes to objects are still visible,
like changes in local lights of texture full brightness.